Leveling VB was too busted even with a nerf to 120/120 -20/20 per kill
